Mass Vaccuum working with one sizable bridge added. Weapon kicks ironed out. fixed weapon switching bug

+++ Assets/Scripts/Sizeable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Sizeable : MonoBehaviour
 {
-    public virtual int getAmount() { return 0; }
+    public virtual float getAmount() { return 0; }
     public virtual bool grow() { return false; }
     public virtual bool shrink() { return false; }
 }
